requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
whisper==1.1.10